locfig fix (new variables: fieldid, spicyid, name)
fixes and uncomments the code to present the locfig as part of the graph

This code requires you have a folder in your main directory called /figures/ with the location images of each of your sources in it. To do this automatically, run the Figure Getter notebook I also uploaded. 

It also requires, in order to pull the images, that you have the following variables defined somewhere prior:
    fieldid = 'G328' (ex. - whatever your region is)
    spicyid = 31415 (ex. - whatever source you're looking at)
    name = 'btingle' (however your name appears in your directory, aka /home/yourname)

I also tweaked it so that it doesn't attempt to display the chi2limit (as in chi2limit, not deltachi2limit) since I suspect that we each have that defined in different ways or not at all. If anyone wants to reintroduce the chi2 label that'd be welcome.