Adding SPICY source notebook
This notebook can be run for any source by changing the sourceindex according to the .txt file. Important note: in order to generate tables for a source with some index i, the notebook must first be run for the sources with indices 0 to i-1 since the table is built sequentially.