removal of deltachi2limit variable
deltachi2limit is an arbitrarily-selected number (say, 200 or so) that was initially meant to be "how far above the lowest chi2 value will we set our limit?" We since changed the criteria to be 3x the lowest chi2limit, but have continued (or at least I have) to use the deltachi2limit variable, sometimes defining it to be equal to chi2limit, which creates inconsistencies. So these changes take a chi2limit variable (which needs to be defined outside of plot_fit.py) and uses that.