Update table_loading.py
similar to the last update, magcols/emagcols can't be defined as defaults to be imported anymore because it depends on whether it's a UKIDSS or VIRAC source. i have a process where i separate the two, run mag_to_flux on each of them using the correct magcols/emagcols, then recombine them.